deleting unused code, cleaning up controllers a bit
